Added StartPage search query
Startpage.com search query does not always have the search in the URL parameter. The next best option is to get the value from the search element.
Also added an error for any status that isn't 200, to show it was not working. I was having issues because it told me it was waiting for ChatGPT to answer, but it had already thrown an error. This at least shows that something is wrong.